// src/Types/Product.ts
export interface SubAttributes {
  _id: string;
  name: string;
  colorCode?: string;
}

export interface Attributes {
  colorIds: SubAttributes[];
  sizeIds: SubAttributes[];
  materialIds: SubAttributes[];
  fabricIds: SubAttributes[];
  occasionIds: SubAttributes[];
}

export interface SEO {
  keywords: string[];
}

export interface colorImages {
  colorId: {
    colorCode: string;
    name: string;
    _id: string;
  };
  images: string[];
  _id: string;
}

export interface ProductType {
  _id: string;
  name: string;
  slug: string;
  description: string;
  price: number;
  salePrice: number;
  sku: string;
  images: string[];
  categoryId: {
    _id: string;
    name: string;
  };
  uniqueCategoryId: {
    _id: string;
    name: string;
  };
  tags: string[];
  attributes: Attributes;
  colorImages?: colorImages[];
  frontImage: string;
  backImage: string;
  stock: number;
  isNewArrival: boolean;
  isBestSelling: boolean;
  isFeatured: boolean;
  showOnHomepage: boolean;
  seo: SEO;
  rating: number;
  isDeleted: boolean;
  isBlocked: boolean;
  reviews: any[];
  createdAt: string;
  updatedAt: string;
  isOffer: boolean;
  offerPrice: number;
  weight: number;
}